Split Binance streaming service implementations apart
